// src/entities/CreditPayment/domain/interfaces/ICreditPaymentController.ts
import type { Request, Response } from "express";

export interface ICreditPaymentController {
  get: (req: Request, res: Response) => Promise<void>;
  getAll: (req: Request, res: Response) => Promise<void>;
  getAllByUser: (req: Request, res: Response) => Promise<void>;
  post: (req: Request, res: Response) => Promise<void>;
  patch: (req: Request, res: Response) => Promise<void>;
  delete: (req: Request, res: Response) => Promise<void>;
}